Move lazy allocator into allocator package
Place LazyMemoryAllocator under lmcache.v1.memory_allocators while keeping the old module path as a compatibility shim.
